requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
Pillow>=10.0
python-decouple>=3.8

# Development
django-extensions>=3.2
ipython>=8.0

# Production (optional)
# gunicorn>=21.0
# psycopg2-binary>=2.9   # PostgreSQL adapter
# django-storages>=1.14  # S3 / cloud media
# boto3>=1.34             # AWS SDK
